// src/ifcopenshell-ts/src/generated/ifc4x3.d.ts
export type IfcChangeActionEnum = string;
// ...
export type IfcStateEnum = string;
// ...
export type IfcWallTypeEnum = string;
// ...
export interface IfcApplication {
	ApplicationDeveloper: IfcOrganization;
	Version: string;
	ApplicationFullName: string;
	ApplicationIdentifier: string;
}
// ...
export interface IfcObjectPlacement {
	PlacementRelTo: IfcObjectPlacement | null;
}
// ...
export interface IfcOwnerHistory {
	OwningUser: IfcPersonAndOrganization;
	OwningApplication: IfcApplication;
	State: IfcStateEnum | null;
	ChangeAction: IfcChangeActionEnum | null;
	LastModifiedDate: number | null;
	LastModifyingUser: IfcPersonAndOrganization | null;
	LastModifyingApplication: IfcApplication | null;
	CreationDate: number;
}
// ...
export interface IfcPersonAndOrganization {
	ThePerson: IfcPerson;
	TheOrganization: IfcOrganization;
	Roles: IfcActorRole[] | null;
}
// ...
export interface IfcProductRepresentation {
	Name: string | null;
	Description: string | null;
	Representations: IfcRepresentation[];
}
// ...
export interface IfcRepresentation {
	ContextOfItems: IfcRepresentationContext;
	RepresentationIdentifier: string | null;
	RepresentationType: string | null;
	Items: IfcRepresentationItem[];
}
// ...
export interface IfcWall {
	GlobalId: string;
	OwnerHistory: IfcOwnerHistory | null;
	Name: string | null;
	Description: string | null;
	ObjectType: string | null;
	ObjectPlacement: IfcObjectPlacement | null;
	Representation: IfcProductRepresentation | null;
	Tag: string | null;
	PredefinedType: IfcWallTypeEnum | null;
}